+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 ## NEXT
 ***
+**🔧 Fixes**
+* Improve exceptions for redundant parentheses in multi-line cases + clarify documentation in that area ([#25](https://github.com/robsdedude/flake8-picky-parentheses/pull/25)).
 
 
 ## 0.3.0